+++ crates/ptool-lua/src/lua_world.rs
@@ -619,6 +619,10 @@ impl LuaWorld {
         crate::redis::connect(value, &self.engine)
     }
 
+    pub(crate) fn s3_connect(&self, options: Table) -> mlua::Result<crate::s3::LuaS3Connection> {
+        crate::s3::connect(options, &self.engine)
+    }
+
     pub(crate) fn git_open(&self, path: Option<String>) -> mlua::Result<crate::git::LuaGitRepo> {
         crate::git::open(path, self.current_dir(), &self.engine)
     }